backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
asyncpg==0.30.0
pgvector==0.3.6
pydantic==2.10.2
python-dotenv==1.0.1
langchain-text-splitters==1.1.0
httpx==0.28.1